Moved HDFS resource stream classes to separate files.
